Stylistic changes and generic message
