+++ sdk/honcho/__init__.py
@@ -1,3 +1,4 @@
 from .client import AsyncClient
-from .syncclient import Client
+from .sync_client import Client
+from .schemas import Message
 from .cache import LRUCache
